tests/wpt/web-platform-tests/css/css-writing-modes-3/text-orientation-parsing-sideways-right-001.html

<!DOCTYPE html>
<title>CSS Writing Modes: parsing text-orientation</title>
<link rel="author" title="Koji Ishii" href="mailto:kojiishi@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#text-orientation">
<meta name="assert" content="This test asserts the parser and getComputedStyle works correctly for the text-orientation property.">
<meta name="flags" content="dom may">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div style="text-orientation: sideways-right"
  data-expected="sideways"></div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).textOrientation;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
